Base de datos, modelos y relaciones de entidades.
Modelos: Se realizan todas las entidades con sus interfaces, tipos, roles, etc.

Relaciones: Se realizan las relaciones entre las entidades.

Base de datos y sincronización: Se realiza conexión a PostgreSQL y sincronización de modelos definidos al inicializar la aplicación. Manejo de errores durante la sincronización y cierre de la conexión  a la base de datos.
