+++ backend/src/config/sync.db.ts
@@ -2,12 +2,12 @@ import PostgreDB from "./db";
 
 const syncDB = async () => {
   try {
-    await PostgreDB.getInstance().sync();
+    const dbInstance = await PostgreDB.getInstance()
+    await dbInstance.sync();
     console.log('Sync database tables with Sequelize...');
+    await dbInstance.close();
   } catch (err) {
     console.error('Unable to sync the database:', err);
-  } finally {
-    PostgreDB.getInstance().close();
   }
 }
