+++ backend/src/models/associationmodels/index.ts
@@ -1,0 +1,1 @@
+export * from './association.models';
