+++ backend/src/models/creditcard/creditcard.interface.ts
@@ -1,0 +1,8 @@
+export interface CreditCardAttributes {
+    id: number;
+    number: string;
+    cvc: string;
+    exp_month: string;
+    exp_year: string;
+    user_id: string;
+}
