+++ backend/src/models/creditcard/index.ts
@@ -1,0 +1,2 @@
+export * from './creditcard.interface';
+export * from './creditcard.model';
