+++ backend/src/models/event/event.status.ts
@@ -1,0 +1,6 @@
+export enum EventStatus {
+    SCHEDULED = 'scheduled',
+    ONGOING = 'ongoing',
+    COMPLETED = 'completed',
+    CANCELLED = 'cancelled',
+}
