+++ backend/src/models/event/event.type.ts
@@ -1,0 +1,4 @@
+export enum EventType {
+    FREE = 'free',
+    PAID = 'paid',
+}
