+++ backend/src/models/event/index.ts
@@ -1,0 +1,4 @@
+export * from './event.interface';
+export * from './event.model';
+export * from './event.status';
+export * from './event.type';
