+++ backend/src/models/guest/guest.interface.ts
@@ -1,0 +1,9 @@
+export interface GuestAttributes {
+    id: number;
+    fullname: string;
+    description: string;
+    user_id: string;
+    event_id: string;
+    email: string;
+    phone: string;
+}
