+++ backend/src/models/guest/index.ts
@@ -1,0 +1,2 @@
+export * from './guest.interface';
+export * from './guest.model';
