+++ backend/src/models/index.ts
@@ -1,0 +1,7 @@
+export * from './creditcard'
+export * from './event'
+export * from './guest'
+export * from './invitation'
+export * from './subscriptiontype'
+export * from './user'
+export * from './associationmodels'
