+++ backend/src/models/invitation/index.ts
@@ -1,0 +1,4 @@
+export * from './invitation.interface';
+export * from './invitation.model';
+export * from './invitation.status';
+export * from './invitation.type';
