+++ backend/src/models/invitation/invitation.status.ts
@@ -1,0 +1,5 @@
+export enum InvitationStatus {
+    ACCEPTED = 'accepted',
+    PENDING = 'pending',
+    REJECTED = 'rejected',
+}
