+++ backend/src/models/invitation/invitation.type.ts
@@ -1,0 +1,4 @@
+export enum InvitationType {
+    FREE = 'free',
+    PAY = 'pay',
+}
