+++ backend/src/models/subscriptiontype/index.ts
@@ -1,0 +1,3 @@
+export * from './subscriptiontype.enum';
+export * from './subscriptiontype.interface';
+export * from './subscriptiontype.model';
