+++ backend/src/models/subscriptiontype/subscriptiontype.enum.ts
@@ -1,0 +1,4 @@
+export enum SubscriptionTypeEnum {
+    FREE = 'free',
+    PREMIUM = 'premium',
+}
