+++ backend/src/models/subscriptiontype/subscriptiontype.interface.ts
@@ -1,0 +1,8 @@
+import { SubscriptionTypeEnum } from "./subscriptiontype.enum";
+
+export interface SubscriptionTypeAttributes {
+    id: number;
+    type: SubscriptionTypeEnum;
+    max_events: number;
+    price: number;
+}
