+++ backend/src/models/subscriptiontype/subscriptiontype.model.ts
@@ -1,17 +1,7 @@
 import { DataTypes, Model, Optional } from 'sequelize';
-import { sequelize } from '../config/db'
-
-enum SubscriptionTypeEnum {
-  FREE = 'free',
-  PREMIUM = 'premium',
-}
-
-export interface SubscriptionTypeAttributes {
-  id: number;
-  type: SubscriptionTypeEnum;
-  max_events: number;
-  price: number;
-}
+import { sequelize } from '../../config/db'
+import { SubscriptionTypeEnum } from './subscriptiontype.enum';
+import { SubscriptionTypeAttributes } from './subscriptiontype.interface';
 
 interface SubscriptionTypeCreationAttributes extends Optional<SubscriptionTypeAttributes, 'id'> { }
