+++ backend/src/models/user/index.ts
@@ -1,0 +1,3 @@
+export * from './user.interface';
+export * from './user.model';
+export * from './user.role';
