+++ backend/src/models/user/user.interface.ts
@@ -1,0 +1,16 @@
+import { UserRole } from "./user.role";
+
+export interface UserAttributes {
+    id: string;
+    fullname: string;
+    email: string;
+    password: string;
+    profile_image: string;
+    phone: string;
+    credit_card: string;
+    subscription_type_id: number;
+    role: UserRole;
+    is_active: boolean;
+    created_at: Date;
+    updated_at: Date;
+}
