+++ backend/src/models/user/user.role.ts
@@ -1,0 +1,4 @@
+export enum UserRole {
+    USER = 'user',
+    ADMIN = 'admin',
+}
