+++ backend/src/services/user.service.ts
@@ -1,5 +1,5 @@
 import UserDAO from '../daos/user.dao';
-import { UserAttributes } from '../models/user.model';
+import { UserAttributes } from '../models/user/user.model';
 import AuthHelper from '../helpers/auth.helper';
 //import EmailHelper from '../helpers/email.helper';
